// frontend/src/app/shared/interfaces/colors.interface.ts
export interface IColors {
    id_colors: number;
    primary_color: string;
    secondary_color: string;
    ternary_color: string;
    cuarternary_color: string;
    neutral_color: string;
    is_selected: number;
}

export interface IColorsPreview {
    primaryColor: string;
    secondaryColor: string;
    ternaryColor: string;
    quaternaryColor: string;
    neutralColor: string;
}